Fixes bug with Drip::Response comparison
With SolidusDrip::ShopperActivity methods being called as part
of the state machine callbacks there exists a moment when the
response from the calls within Solidus::Drop::ShopperActivity
were being compared with the boolean `false`.

Drip::Response objects implement their own `#==` method which
expects to be used in comparison with other Drip::Response objects.
In the case of a Drip::Response being compared to `false` this
would throw an error.

Having the SolidusDrip::ShopperActivity methods return a boolean
fixes this issue.

`state_machines-activemodel` attempts to compare the response to false.
https://github.com/state-machines/state_machines-activemodel/blob/master/lib/state_machines/integrations/active_model.rb#L428

`Drip::Response` attempts to compare one response to another.
https://github.com/state-machines/state_machines-activemodel/blob/master/lib/state_machines/integrations/active_model.rb#L428